example: enforce token buckets on folders
I have a script called tb which tracks token buckets.  For example, I
can say "create a token bucket named 'better' which refills at the
rate of 1 token per day".  Running "tb better 1" tries to take a
single token from that bucket.  Only if it succeeds do I proceed to
selecting the given folder.

This helps me avoid checking my email too frequently.

+++ examples/mailz.sh
@@ -333,6 +333,11 @@ unsubscribe_url() {
 
 # select a folder
 select_folder() {
+    # only proceed if this folder's token bucket has enough balance
+    if ! tb $1 1; then
+        return
+    fi
+    
     cd "${MAIL}/$1"
     rm -f "tmp/${message_list}"
     list
